refactor(#10706): share the contact-delete handler and OpenAPI docs
The person and place DELETE endpoints delete a hierarchy the same way, so their
controller handlers and @openapi blocks were near-identical. Move the shared
handler into the delete-contact service (the controllers now just point their
delete at it) and pull the delete_users/dry_run params and the response shapes
into shared OpenAPI components referenced from both.
